Fix choose_your_own rendering
I'm not 100% sure why, but when I run the code without this change, I see only the blue and red points in the visualization, not the coloring of the background to indicate which class my algorithm would assign to an arbitrary point. My best guess is that plt.show() is a blocking method, so it prevents the user's classifier from running until they close the popup.. at which point they cannot see the results.

This change waits until everything is calculated before calling `plt.show()`, and seems to work correctly.